@@ -9,14 +9,6 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
-  "peerDependencies": {
-    "@material-ui/core": "4.11.0",
-    "@material-ui/icons": "4.9.1",
-    "@material-ui/lab": "4.0.0-alpha.58",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "react-router-dom": "6.2.2"
-  },
   "dependencies": {
     "@graasp/ui": "link:..",
     "@testing-library/jest-dom": "link:../node_modules/@testing-library/jest-dom",
@@ -26,19 +18,15 @@
     "@types/node": "link:../node_modules/@types/node",
     "@types/react": "link:../node_modules/@types/react",
     "@types/react-dom": "link:../node_modules/@types/react-dom",
-    "@types/react-router-dom": "link:../node_modules/@types/react-router-dom",
-    "immutable": "4.0.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-router-dom": "6.2.2",
+    "@types/react-router-dom": "5.3.3",
+    "immutable": "4.1.0",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
     "react-scripts": "5.0.1",
     "typescript": "4.7.4"
   },
   "devDependencies": {
-    "@babel/plugin-syntax-object-rest-spread": "^7.8.3",
-    "@material-ui/core": "4.12.4",
-    "@material-ui/icons": "4.11.3",
-    "@material-ui/lab": "4.0.0-alpha.61"
+    "@babel/plugin-syntax-object-rest-spread": "^7.8.3"
   },
   "eslintConfig": {
     "extends": "react-app"
