@@ -54,25 +55,27 @@
     "clsx": "1.1.1",
     "http-status-codes": "2.2.0",
     "immutable": "4.1.0",
-    "katex": "0.15.3",
+    "katex": "0.16.0",
     "lodash.truncate": "4.4.2",
-    "qs": "6.10.3",
+    "qs": "6.10.5",
     "react-cookie-consent": "7.4.1",
-    "react-i18next": "11.16.9",
-    "react-quill": "1.3.5",
+    "react-i18next": "11.17.0",
+    "react-quill": "2.0.0-beta.4",
     "react-rnd": "10.3.7",
     "react-text-mask": "5.4.3",
     "uuid": "8.3.2"
   },
   "peerDependencies": {
-    "@material-ui/core": "4.11.0",
-    "@material-ui/icons": "4.9.1",
-    "@material-ui/lab": "4.0.0-alpha.58",
-    "ag-grid-community": "27.3.0",
-    "ag-grid-react": "27.3.0",
-    "i18next": "21.3.1",
-    "react": "^16.13.1",
-    "react-dom": "16.13.1",
+    "@emotion/react": "11.10.4",
+    "@emotion/styled": "11.10.4",
+    "@mui/icons-material": "5.8.3",
+    "@mui/lab": "~5.0.0-alpha.85",
+    "@mui/material": "5.10.7",
+    "ag-grid-community": "28.1.1",
+    "ag-grid-react": "28.1.1",
+    "i18next": "~21.8.9",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
     "react-router-dom": "6.2.2"
   },
   "peerDependenciesMeta": {
@@ -85,80 +88,85 @@
   },
   "devDependencies": {
     "@babel/core": "^7.18.9",
-    "@babel/plugin-proposal-optional-chaining": "7.16.7",
-    "@babel/preset-env": "7.17.10",
-    "@babel/preset-typescript": "7.16.7",
-    "@commitlint/cli": "17.0.0",
-    "@commitlint/config-conventional": "17.0.0",
-    "@material-ui/core": "4.12.4",
-    "@material-ui/icons": "4.11.3",
-    "@material-ui/lab": "4.0.0-alpha.61",
+    "@babel/plugin-proposal-optional-chaining": "7.17.12",
+    "@babel/preset-env": "7.18.2",
+    "@babel/preset-typescript": "7.17.12",
+    "@commitlint/cli": "17.0.2",
+    "@commitlint/config-conventional": "17.0.2",
+    "@emotion/react": "11.10.4",
+    "@emotion/styled": "11.10.4",
     "@mdx-js/react": "^1.6.22",
+    "@mui/icons-material": "5.10.6",
+    "@mui/lab": "5.0.0-alpha.101",
+    "@mui/material": "5.10.7",
     "@rollup/plugin-babel": "5.3.1",
     "@rollup/plugin-commonjs": "22.0.0",
     "@rollup/plugin-node-resolve": "13.3.0",
     "@rollup/plugin-typescript": "8.3.2",
+    "@storybook/addon-a11y": "6.5.12",
     "@storybook/addon-actions": "^6.5.9",
+    "@storybook/addon-coverage": "0.0.4",
     "@storybook/addon-docs": "6.5.9",
     "@storybook/addon-essentials": "6.5.9",
-    "@storybook/addon-interactions": "^6.5.9",
+    "@storybook/addon-interactions": "6.5.12",
     "@storybook/addon-links": "^6.5.9",
     "@storybook/builder-webpack4": "^6.5.9",
+    "@storybook/jest": "0.0.10",
     "@storybook/manager-webpack4": "^6.5.9",
     "@storybook/react": "^6.5.9",
-    "@storybook/testing-library": "^0.0.13",
+    "@storybook/test-runner": "0.7.2",
+    "@storybook/testing-library": "0.0.13",
     "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.2.0",
+    "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^14.2.0",
-    "@types/enzyme": "3.10.12",
-    "@types/enzyme-adapter-react-16": "1.0.6",
-    "@types/jest": "28.1.4",
+    "@trivago/prettier-plugin-sort-imports": "3.2.0",
+    "@types/jest": "28.1.0",
     "@types/katex": "0.14.0",
-    "@types/lodash.truncate": "4.4.7",
-    "@types/node": "^17.0.33",
+    "@types/node": "^17.0.39",
     "@types/qs": "6.9.7",
-    "@types/react": "17.0.47",
-    "@types/react-dom": "17.0.17",
+    "@types/react": "18.0.10",
+    "@types/react-dom": "18.0.5",
     "@types/react-router-dom": "5.3.3",
     "@types/react-text-mask": "5.4.11",
     "@types/uuid": "8.3.4",
-    "@typescript-eslint/eslint-plugin": "^5.5.0",
-    "@typescript-eslint/parser": "^5.5.0",
-    "ag-grid-community": "27.3.0",
-    "ag-grid-react": "27.3.0",
+    "@typescript-eslint/eslint-plugin": "^5.27.0",
+    "@typescript-eslint/parser": "5.27.0",
+    "ag-grid-community": "28.1.1",
+    "ag-grid-react": "28.1.1",
     "babel-jest": "28.1.0",
     "babel-loader": "^8.2.5",
+    "concurrently": "7.4.0",
     "cross-env": "7.0.3",
-    "eslint": "7.32.0",
+    "eslint": "^8.16.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-config-standard": "17.0.0",
-    "eslint-config-standard-react": "11.0.1",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-n": "15.2.1",
+    "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-react": "^7.30.1",
+    "http-server": "14.1.1",
     "husky": "^8.0.1",
     "jest": "28.1.0",
-    "jest-environment-jsdom": "28.1.2",
-    "node-sass": "7.0.1",
     "npm-run-all": "4.1.5",
     "pinst": "^3.0.0",
-    "prettier": "^2.7.1",
-    "react": "^16.13.1",
-    "react-dom": "16.13.1",
-    "react-query": "3.39.0",
-    "react-test-renderer": "17.0.2",
+    "prettier": "^2.6.2",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
+    "react-query": "3.39.1",
+    "react-router-dom": "6.2.2",
+    "react-test-renderer": "18.1.0",
     "rimraf": "3.0.2",
-    "rollup": "2.73.0",
+    "rollup": "2.75.5",
     "rollup-jest": "2.0.0",
     "rollup-plugin-copy": "3.4.0",
     "rollup-plugin-peer-deps-external": "2.2.4",
     "rollup-plugin-scss": "3.0.0",
+    "sass": "1.52.2",
     "standard-version": "9.5.0",
     "storybook": "6.5.9",
     "ts-jest": "28.0.5",
-    "ts-node": "10.7.0",
-    "typescript": "4.6.4"
+    "ts-node": "10.8.0",
+    "typescript": "4.7.2",
+    "wait-on": "6.0.1"
   },
   "files": [
     "dist"
