@@ -41,10 +41,11 @@ export default {
     }),
   ],
   external: [
-    '@material-ui/core',
-    '@material-ui/icons',
-    '@material-ui/lab',
-    '@material-ui/styles',
+    '@mui/material',
+    '@mui/icons-material',
+    '@mui/lab',
+    '@emotion/styled',
+    '@emotion/react',
     'ag-grid-community',
     'ag-grid-react',
     'katex',
