@@ -1,53 +1,57 @@
-// make it a module
 export {};
-// import React, { Component, ReactNode } from 'react';
-// import { v4 } from 'uuid';
+// import { StatusCodes, getReasonPhrase } from 'http-status-codes';
 // import { Map } from 'immutable';
-// import { shallow, ShallowWrapper } from 'enzyme';
-// import Alert from '@material-ui/lab/Alert';
+// import { v4 } from 'uuid';
+
+// import Alert from '@mui/material/Alert';
+
+// import React, { Component, ReactNode } from 'react';
+
 // import Loader from '../Loader';
+// import { FORBIDDEN_TEXT, SETTINGS } from '../constants';
+// import ForbiddenContent from './ForbiddenContent';
 // import ItemLoginAuthorization, {
 //   ItemLoginAuthorizationProps,
 // } from './ItemLoginAuthorization';
-// import { getReasonPhrase, StatusCodes } from 'http-status-codes';
-// import { SETTINGS } from '../constants';
 // import ItemLoginScreen from './ItemLoginScreen';
-//
+
 // const defaultData: {
-//   data?: any;
+//   data?: unknown;
 //   isLoading?: boolean;
 //   isError?: boolean;
-//   error?: { message: any };
+//   error?: { message: unknown };
 // } = {
 //   data: null,
 //   isLoading: false,
 //   isError: false,
 //   error: undefined,
 // };
-//
+
 // const createProps = ({
 //   itemId = '',
 //   useCurrentMemberData = defaultData,
 //   useItemData = defaultData,
 //   useItemLoginData = defaultData,
+//   signOut = jest.fn(),
 //   signIn = jest.fn(),
 // }): ItemLoginAuthorizationProps => ({
 //   useCurrentMember: jest.fn().mockReturnValue(useCurrentMemberData),
 //   useItem: jest.fn().mockReturnValue(useItemData),
 //   useItemLogin: jest.fn().mockReturnValue(useItemLoginData),
 //   itemId,
+//   signOut,
 //   signIn,
 // });
-//
+
 // class MockChildren extends Component {
 //   render(): ReactNode {
 //     return <div />;
 //   }
 // }
-//
+
 // describe('ItemLoginAuthorization', () => {
 //   let wrapper: ShallowWrapper;
-//
+
 //   it('renders correctly', () => {
 //     const props = createProps({
 //       itemId: v4(),
@@ -57,7 +61,7 @@ export {};
 //     wrapper = shallow(<Comp />);
 //     expect(wrapper).toMatchSnapshot();
 //   });
-//
+
 //   it('renders <Loader /> if member is loading', () => {
 //     const props = createProps({
 //       itemId: v4(),
@@ -68,7 +72,7 @@ export {};
 //     const loader = wrapper.find(Loader);
 //     expect(loader).toHaveLength(1);
 //   });
-//
+
 //   it('renders <Loader /> if item is loading', () => {
 //     const props = createProps({
 //       itemId: v4(),
@@ -142,5 +146,26 @@ export {};
 //     wrapper = shallow(<Comp />);
 //     const loader = wrapper.find(ItemLoginScreen);
 //     expect(loader).toHaveLength(1);
+//   });
+//   it('renders ForbiddenContent if user is defined and item is undefined', () => {
+//     const props = createProps({
+//       itemId: v4(),
+//       useCurrentMemberData: { data: Map({ id: v4() }) },
+//     });
+//     const Comp = ItemLoginAuthorization(props)(MockChildren);
+//     wrapper = shallow(<Comp />);
+//     const loader = wrapper.find(ForbiddenContent);
+//     expect(loader).toHaveLength(1);
+//   });
+//   it('renders ForbiddenContent if item login is undefined and cannot fetch item', () => {
+//     const props = createProps({
+//       itemId: v4(),
+//       useItemLoginData: { data: Map({ id: v4() }) },
+//       useItemData: { isError: true, error: { message: 'error' } },
+//     });
+//     const Comp = ItemLoginAuthorization(props)(MockChildren);
+//     wrapper = shallow(<Comp />);
+
+//     expect(wrapper.html()).toContain(FORBIDDEN_TEXT);
 //   });
 // });
