@@ -1,6 +1,8 @@
-import React from 'react';
-import { Main, MainMenu, MenuItem } from '@graasp/ui';
 import HomeIcon from '@material-ui/icons/Home';
+
+import React from 'react';
+
+import { Main, MainMenu, MenuItem } from '@graasp/ui';
 import '@graasp/ui/dist/bundle.css';
 
 const App = () => {
