@@ -1,9 +1,12 @@
-import React from 'react';
 import { RecordOf } from 'immutable';
-import Collapse from './Collapse';
-import type { Item, UnknownExtra } from '../types';
 
-interface WithCollapseProps<T> {
+import React from 'react';
+
+import type { Item, UnknownExtra } from '@graasp/sdk';
+
+import Collapse from './Collapse';
+
+interface WithCollapseProps<T extends UnknownExtra> {
   item: RecordOf<Item<T>>;
 }
 
@@ -11,7 +14,7 @@ function withCollapse<T extends UnknownExtra>({ item }: WithCollapseProps<T>) {
   return (component: JSX.Element): JSX.Element => {
     class ComponentWithCollapse extends React.Component {
       render(): JSX.Element {
-        return <Collapse title={item.name} content={component} />;
+        return <Collapse title={item.name}>{component}</Collapse>;
       }
     }
 
