@@ -1,7 +1,13 @@
+import DragIndicatorIcon from '@mui/icons-material/DragIndicator';
+import { Box, styled } from '@mui/material';
+
 import React, { FC, useEffect, useRef } from 'react';
-import DragIndicatorIcon from '@material-ui/icons/DragIndicator';
-import { makeStyles } from '@material-ui/core/styles';
-import { DRAG_ICON_SIZE } from '../constants';
+
+const StyledDragIndicatorIcon = styled(DragIndicatorIcon)({
+  '&:hover': {
+    cursor: 'move',
+  },
+});
 
 interface Props {
   data: {
@@ -11,40 +17,27 @@ interface Props {
   buildId?: (id: string) => string;
 }
 
-const useStyles = makeStyles(() => ({
-  dragIconContainer: {
-    display: 'flex',
-    '&:hover': {
-      cursor: 'move',
-    },
-  },
-  dragIcon: {
-    fontSize: DRAG_ICON_SIZE,
-  },
-}));
-
 const DragCellRenderer: FC<Props> = ({
   data: item,
   registerRowDragger,
   buildId,
 }) => {
-  const classes = useStyles();
   const dragRef = useRef<SVGSVGElement>(null);
 
   useEffect(() => {
     if (dragRef && dragRef.current) {
       registerRowDragger(dragRef.current);
     }
-  }, [dragRef]);
+  }, [dragRef, registerRowDragger]);
 
   return (
-    <div className={classes.dragIconContainer}>
-      <DragIndicatorIcon
+    <Box display='flex' alignItems='center'>
+      <StyledDragIndicatorIcon
+        fontSize='small'
         id={buildId?.(item.id)}
         ref={dragRef}
-        className={classes.dragIcon}
       />
-    </div>
+    </Box>
   );
 };
 
