@@ -1,58 +1,47 @@
+import { styled } from '@mui/material';
+import Toolbar from '@mui/material/Toolbar';
+import Typography from '@mui/material/Typography';
+
 import React, { FC } from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-import clsx from 'clsx';
 
 interface Props {
   selected: string[];
   Actions?: React.FC<{ selectedIds: string[] }>;
   NoSelectionToolbar?: React.FC;
-  countText?: string;
+  countTextFunction?: (selection: string[]) => string;
 }
 
-const useToolbarStyles = makeStyles((theme) => ({
-  root: {
-    paddingLeft: theme.spacing(2),
-    paddingRight: theme.spacing(1),
-  },
-  title: {
-    flex: '1 1 100%',
-    display: 'flex',
-    alignItems: 'center',
-  },
-  highlight: {
-    background: theme.palette.primary.main,
-    color: 'white',
-  },
-}));
+const StyledTitle = styled(Typography)({
+  flex: '1 1 100%',
+  display: 'flex',
+  alignItems: 'center',
+});
 
 const TableToolbar: FC<Props> = ({
   selected,
   Actions,
   NoSelectionToolbar,
-  countText,
+  countTextFunction,
 }) => {
-  const classes = useToolbarStyles();
   const numSelected = selected.length;
+  const StyledToolbar = styled(Toolbar)(({ theme }) => ({
+    paddingLeft: theme.spacing(2),
+    paddingRight: theme.spacing(1),
+    ...(numSelected > 0 && {
+      background: theme.palette.primary.main,
+      color: 'white',
+    }),
+  }));
 
   if (numSelected > 0) {
     return (
-      <Toolbar
-        className={clsx(classes.root, {
-          [classes.highlight]: numSelected > 0,
-        })}
-      >
-        <Typography
-          className={classes.title}
-          color='inherit'
-          variant='subtitle1'
-          component='div'
-        >
-          {countText ?? `${numSelected} selected`}
-        </Typography>
+      <StyledToolbar>
+        {/* @ts-expect-error: component does exist on typography */}
+        <StyledTitle color='inherit' variant='subtitle1' component='div'>
+          {countTextFunction?.(selected) ?? `${numSelected} selected`}
+        </StyledTitle>
         {Actions?.({ selectedIds: selected })}
-      </Toolbar>
+      </StyledToolbar>
     );
   }
 
