@@ -1,4 +1,6 @@
-export { default as AnalyzeIcon } from './AnalyzeIcon';
-export { default as ExploreIcon } from './ExploreIcon';
+export { default as AnalyticsIcon } from './AnalyticsIcon';
+export { default as LibraryIcon } from './LibraryIcon';
 export { default as PlayIcon } from './PlayIcon';
 export { default as BuildIcon } from './BuildIcon';
+export { default as ItemIcon } from './ItemIcon';
+export { default as CCLicenseIcon } from './CCLicenseIcon';
