@@ -1,12 +1,35 @@
-import React from 'react';
+import { styled } from '@mui/material';
+
+import React, { FC, useEffect, useState } from 'react';
 import { Rnd } from 'react-rnd';
+
+import {
+  getIframeResizeHeightCookie,
+  setIframeResizeHeightCookie,
+} from '@graasp/sdk';
+
+import { IFRAME_MIN_HEIGHT, ITEM_MAX_HEIGHT } from '../constants';
 import ResizingIcon from '../icons/ResizingIcon';
 import { UUID } from '../types';
-import { getIframeResizeHeightCookie, setIframeResizeHeightCookie } from '@graasp/sdk';
 
-interface WithResizingProps {
+export const StyledIFrame = styled('iframe')<{
+  isResizable?: boolean;
   height: string | number;
-  memberId: UUID;
+}>(({ isResizable, height }) =>
+  isResizable
+    ? {
+        width: '100%',
+        border: 'none',
+        maxHeight: !isResizable ? ITEM_MAX_HEIGHT : undefined,
+        height: !isResizable ? height : '100%',
+      }
+    : {},
+);
+
+export interface WithResizingProps {
+  height: string | number;
+  component: JSX.Element;
+  memberId?: UUID;
   itemId: UUID;
 }
 
@@ -31,65 +54,47 @@ const resizeHandleStyles = {
   },
 };
 
-function withResizing({ height, memberId, itemId }: WithResizingProps) {
-  return (component: JSX.Element): JSX.Element => {
-    class ComponentWithResizing extends React.Component<
-      {},
-      { variableHeight: string | number }
-    > {
-      constructor(props: any) {
-        super(props);
-        this.state = {
-          variableHeight: height,
-        };
-      }
+const withResizing =
+  <P extends object>({
+    height,
+    component,
+    memberId,
+    itemId,
+  }: WithResizingProps): FC<P> =>
+  () => {
+    const [variableHeight, setVariableHeight] = useState<number | string>(
+      getIframeResizeHeightCookie({ memberId, itemId }) ?? height,
+    );
 
-      componentDidMount(): void {
-        const iframeResizeHeight = getIframeResizeHeightCookie(memberId, itemId);
+    useEffect(() => {
+      setIframeResizeHeightCookie({ memberId, itemId }, variableHeight);
+    }, [variableHeight]);
 
-        if (iframeResizeHeight) {
-          this.setState({
-            variableHeight: iframeResizeHeight,
-          });
-        }
-      }
-
-      componentDidUpdate(_prevProps: any, prevState: any) {
-        if (prevState.variableHeight !== this.state.variableHeight) {
-          setIframeResizeHeightCookie(memberId, itemId, this.state.variableHeight);
-        }
-      }
-
-      render(): JSX.Element {
-        return (
-          <>
-            <div style={resizeHandleStyles.resizableContainer}>
-              <Rnd
-                size={{ width: '100%', height: this.state.variableHeight }}
-                position={{ x: 0, y: 0 }}
-                style={{ position: 'relative' }}
-                disableDragging
-                enableResizing={{ bottom: true }}
-                resizeHandleComponent={{ bottom: <ResizingIcon /> }}
-                resizeHandleStyles={{
-                  bottom: resizeHandleStyles.resizeHandleComponent,
-                }}
-                onResizeStop={(_e, _direction, ref, _delta, _position) => {
-                  this.setState({
-                    variableHeight: ref.style.height,
-                  });
-                }}
-              >
-                {component}
-              </Rnd>
-            </div>
-          </>
-        );
-      }
-    }
-
-    return <ComponentWithResizing />;
+    return (
+      <div style={resizeHandleStyles.resizableContainer}>
+        <Rnd
+          style={{ position: 'relative' }}
+          disableDragging
+          enableResizing={{ bottom: true }}
+          default={{
+            width: '100%',
+            height: variableHeight,
+            x: 0,
+            y: 0,
+          }}
+          minHeight={IFRAME_MIN_HEIGHT}
+          resizeHandleComponent={{ bottom: <ResizingIcon /> }}
+          resizeHandleStyles={{
+            bottom: resizeHandleStyles.resizeHandleComponent,
+          }}
+          onResizeStop={(_e, _direction, ref) => {
+            setVariableHeight(ref.style.height);
+          }}
+        >
+          {component}
+        </Rnd>
+      </div>
+    );
   };
-}
 
 export default withResizing;
